% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBEPGraphics.R
\name{cbep_colors}
\alias{cbep_colors}
\alias{cbep_colors2}
\title{Default Colors for CBEP Graphics}
\usage{
cbep_colors()

cbep_colors2()
}
\value{
A vector containg six colors.
}
\description{
These functions define default colors for CBEP 2020 State of the Bay Graphics.
Each defines a six color sequence useful for standardizing look and feel of
CBEP graphics.  These are categorical, not continuous color sequences, so they
are not well suited to depicting continuous values.
}
\details{
\code{cbep_colors} returns colors so that adjacent colors contrast nicely.

 \code{cbep_colors2} returns colors from darkest to lightest.
}
\section{Functions}{
\itemize{
\item \code{cbep_colors2}: Alternate sequence of CBEP colors, ordered from darkest to lightest.
}}

\examples{
    cbep_colors()
    cbep_colors()[5]
    
}
