% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBEPGraphics.R
\name{load_cbep_fonts}
\alias{load_cbep_fonts}
\title{CBEP Preferred Font}
\usage{
load_cbep_fonts()
}
\description{
Set up R to use CBEP preferred font family.
}
\details{
This function loads a CBEP preferred font family for graphics.
The function will try to import related fonts, reporting an error if no
members of the font family can be imported. The function loads related
fonts to make them available for use in R. Finally, it sets a default
font for graphical annotations in ggplot.

The current CBEP Graphics font family is Montserrat.
}
\examples{
    load_cbep_fonts()

}
